runtime: Slightly simplify UTF-8 decoding
We don't need to mask out the header byte to determine how long its byte
sequence is, a direct range check is sufficient. This also lets us get
rid of the *ByteHeaderMask constants

On a modern desktop Intel processor this results in less code being
generated but doesn't affect execution speed. This is likely due to
there being enough integer units available to calculate the AND in time
for the next byte to be read from memory.