Initial support for declaring static libraries
Currently all symbols are expected to either come from the system
libraries or liblliby.a. This presents two major issues:

- Every Llambda program links the entire runtime and standard library
  which is quite large, particularily the Unicode character database

- This precludes from linking to libraries outside the stdlib once
  generalised libraries are supported

This adds a new NativeFunction bound value. These can be declared using
(define-native-library). Currently only (static-library) and
system-library are supported. system-library is intended to be a
placeholder that refers to the system standard library and lliby core.

(native-function) now takes a mandatory library as its first parameter.
All the existing functions have been switched to use system-library
except for (scheme time) which has experimentally been separated to its
own library as a proof of concept.

The next commits split out the Llambda runtime in to separate static
libraries where possible.