runtime: Split in to separate static libraries
This separates out the runtime in two classes of static libraries:

- llcore contains the bindings, allocator, core support functions,
  utility functions, etc. This is required by all Llambda libraries and
  expected to exist by codegen itself

- Each Scheme library has its own stdlib library: llbase, llread,
  llwrite, llinexact, etc.

At the same time each library's symbols have been re-prefixed with their
corresponding library name. The name "lliby" survives mainly as the name
of the top-level C++ namespace.

As part of this the Unicode Character Database functionality has been
moved to a separate ucd/ directory. StringCell and UnicodeChar have been
generalised to accept character conversion functions passed in to them
instead of calling out the UCD functions. This allows the UCD to become
part of llchar and avoids linking ~80KiB of Unicode data with every
program

The result of this is an over 1/3rd reduction in size for a simple
program using (scheme base) and (scheme write) on x86-64 Linux with a
release build. This also has the side effect of speeding up the
functional tests due to less data being handled by the LLVM toolchain.