compiler: Default to -O 2 when run from CLI
Unlike a C/C++ compiler Llambda will generate code that's extremely
inefficient compared to -O 2. This is due to much of the language
being built on top of high level constructs such as self-executing
lambdas.

-O 0 is intended mostly for use within functional tests where it
exercises code paths not normally hit by optimised code. It's not a
reasonable default for the command line compiler to use.