compiler: Make ps.WorldPtrValue a singleton
An instance of ps.WorldPtrValue was created for each function to
represent the world pointer argument for that function. This value is
passed through the entire planner and tracked in PlannerState and
PlannedFunction.

The intent was to use it as a canary value to ensure the caller actually
had a world pointer. However, functions typically always start planning
with a world pointer and then remove it at the end of its not needed.
Passing around the world pointer doesn't prevent it from being
erroneously removed at the end of planning. Also, every allocating step
took an implicit world pointer that was only actualised much later in
PlanCellAllocations

Instead make ps.WorldPtrValue a singleton that means "world pointer for
the current function". Steps that need the world pointer now take it
implicitly; it appears in their inputValues but not in their argument
list.

This simplifies the planner and reduces the size of some large
signatures. Minimal changes were required for codegen as it still treats
the world pointer as a normal temp value.