compiler: Support for inlining with captured vars
This allows captured variables to be loaded from a closure for the
purposes of inlining. The closure's structure is persisted in a new
LambdaClosure case class and closure loading is factored out in to a
LoadClosureData function. AttemptInlineApply will look for variables
that it can load from the closure and use LoadClosureData to load them
if possible.

This is generally useful but it's primarily part of groundwork for
polymorphic lambdas. The upper bound instance will build the initial
closure and the specialised instances will load variables from that.