compiler: Make RetypeLambdaArgs polymorph aware
RetypeLambdaArgs was starting with the types the argument's storage
locations which correspond to the upper bound types. Instead use the
types from the specific polymorph being planned.

This prevents unexpected behaviour when dealing with types with
contravariant components such as procedures. For example, the procedure
type from the upper bound was selected if the procedures fixed arguments
were parameterized.