compiler: Avoid unioning identical types
SchemeType.+ was unconditionally unioning the two passed types together.
This is semantically harmless but it will unroll recursive unions even
if they're the same type. This causes it to produce ugly types for
proper lists if they're ever unioned together

Instead, short circuit + to just return "this" if both types are
identical.