compiler: Don't check if rest list is proper
With mutable pairs we were using TempValueToIntermediate to build the
incoming rest argument in PlanProcedureTrampoline. This would convert
the rest argument type to <list-element> which would cause the list's
properness to be checked again when passing to the target procedure.

We know the rest list cannot be modified inside the trampoline so
directly create an iv.CellValue of the correct type.