compiler: Fix ValuesToList length hint
ValuesToList was calculating its list length hint with the assumption
that the passed terminator was an empty list. However, if it was a list
itself this would calculate the wrong as the length of the tail list
wouldn't be included.

Fix to only pass a length hint if the terminator is a known list
element.