Introduce invalid-argument-error?
This is a bit of a dumping ground for errors without more specific
predicates. Fortunately, very few errors fall in to this category. Most
of them are related to illegal operations on ports (input on an output
port, using a closed port, etc).

The main advantage of giving these an error predicate is to further work
on deprecating the overly general expect-failure from the test framework