compiler: Fix exception on (parameterize)
We were annotating llcore_dynamicenv_set_value with nounwind which
means LLVM was allowed to assume it never throws an exception. However,
in the case of attempting to parameterize a non-parameter we will throw
a Scheme exception. At -O 2 LLVM was removing our GC cleanup code which
meant we could crash the next time we enter the garbage collector.