Handle integer overflow in (/)
We're allowed to overflow to inexact in (/) as the result is inexact for
most of its inputs. However, we were computing the wrong value on the
Scala side and on Mac OS X x86-64 we were raising a floating point
exception in the runtime.

Handle by explicitly catching this case and producing an inexact result

Closes #20