compiler: Support for (define) in REPL
This will notice (define) and have the extractor parse it in a body
context and then insert code to print the bound value to stdout. If
printing program succeeds the binding expressions are prefixed to all
future programs.

This is a bit weird if the initialiser has side effects as it will be
executed at the beginning of every REPL program. In the future we could
check if the result of the binding is a Scheme datum and then replace
the bound value with that datum instead of re-evaluating the expression.