compiler: Support for multiple values in REPL
This adds a new (print-thunk-result) helper procedure in a new
(llambda internal repl) library. This behaves the same as (write) except
it pretty prints multiple values by wrapping them in (values ...)

As this requires wrapping the user's expression in a (lambda) it also
allow more type of body expressions to work as expected in the REPL.
However, unless they have special handling like (define) they won't
persist in to future REPL expression.