compiler: Capture REPL program stderr
We were passing REPL programs' stderr directly to the compiler's stderr.
This works during normal REPL use but causes our unit tests to spew to
stderr which is ugly and could confuse tools.

Instead capture the stderr separately from stdout and print it when we
handle a ReplProcessNonZeroExitException. This means we will lose stderr
if the REPL program exits normally. However, there should be no output
to stderr during normal operation unless the user is explicitly writing
to (current-error-port).