Catch division by exact zero in (/)
From R7RS: "It is an error if any argument of / other than the first is
an exact zero."

This is mostly just extra checks. The only code generation difference is
native code is no longer generated for a flonum numerator and an exact
integer denominator as that may cause a divide by zero. We didn't
generate code for the exact integer / exact integer case so this should
be a fairly marginal difference in practice.