stdlib: Don't use WeakListof for native functions
WeakListof is intended to be used with recursive Scheme procedures where
type safety is guaranteed even if non-lists are passed. For native
functions the runtime is expecting to be passed a proper list and can
cause undefined behaviour if a non-list is passed.

Switch to normal (Listof) for native functions and add functional tests
for improper lists.