Include the size of record types in class map
This will allow cloning for record types that include native types.
Cloning can be accomplished by copying the entire record-like
byte-for-byte and then fixing up the cell pointers to point to the new
world.

This bloats the size of the class map but it's constant except for the
two runtime class types. The main effect will slightly larger executable
sizes.

- Prefix the class map with the size of the record-like type

- Always allocate a class map even if the record-like contains no cell
  offsets

- Rename "offset map" to "class map" to reflect that it contains more
  than offsets now