Very basic actor support
This adds an actor library called (llambda actor) that exports a single
(start-actor) procedure. This accepts a procedure, copies the
procedure's closure in to a new world, and launches the procedure in
that world.

There is no way to communicate with or supervise the child actor. The
unit tests are very basic right now for that reason.