runtime: Preserve (eqv?) when cloning
When we were cloning values we were creating a cloned cell for every
reference to a cell. If the same cell was encountered multiple times it
would be cloned each time. Fix by keeping a hash of cloned values and
returning the existing clone when we encounter a cell we've seen before.