runtime: Add very simple scheduler
This enqueues work on worker threads. This doesn't attempt to manage the
number of worker threads; if all threads are busy then a new one is
launched. If a worker thread is idle for for than 5 seconds it will
exit.

This ports the Finalizer to the new scheduler. This isn't a perfect fit
for the scheduler as it's completely CPU bound and can run in the
background. However, once other users are ported to the scheduler this
means that finalizer work can be mixed in with other work.

This will eventualyl be used to support multiplexing actors on threads.