Introduce ActorContext
This moves the Mailbox and last sender from outside of World to a
dedicated ActorContext object. This means the root thread is no longer
an actor. The lack of unification is a bit disappointing but actors are
about to run with a dedicate receive loop which doesn't make sense for
the root thread.

This also introduces a new no-actor-error? which is raised from
procedures that expect an actor context in a non-actor world.