Expire escape procedures when receive returns
We can't resume continuations captured during another receive loop
because we could have been moved to a different thread. We could check
to see if we're running on the same stack but it's better to
deterministically fail to prevent hard-to-debug issues with sporadic
failures.

Implement this by adding a run sequence number to World and incrementing
it on every call to run(). If we attempt to resume with a mismatched
sequence number we will raise an expired-escape-procedure-error.