runtime: Fix race with sleeping actors
We were releasing the mailbox lock after popping the last message from
the stack and re-acquiring it before putting the actor to sleep on the
mailbox. It's possible for another world to enqueue a message while the
lock was released which will cause us to miss the wakeup.

Fix by putting the world to sleep while the lock is still held in
receive(). This should also be more efficient as we avoid a lock/unlock
operation.