runtime: Allocate GC memory with calloc()
When GC debugging is disabled use calloc() to allocate memory blocks.
This is so we can allocate smaller blocks when sending actor messages
without having multiple allocator code paths.

When GC debugging is enabled we still use mmap() so we can invalidate
the memory block after its freed

This also removes the Mac OS X valgrind workaround. This is no longer
required with Valgrind on Yosemite.