compiler: Make dynamic-{true,false} procedures
Because we proxy through a mutable to create these dynamic values the
compiler can't optimise out their initialisation. The result is we
unconditionally allocate two mutable cells on every test. Only a small
handful of tests end up actually using the values.

This has to remove a line from LambdaSignatureSuite that was testing for
input argument types. The compiler can't look in to (dynamic-true) so
it's not sure if it can terminate which means it can't feed in type
information past the point of the call.