compiler: Rework record-like cell allocation
Previously a basic record-like cell would be constructed with
InitRecordLike and then initialised with SetRecordDataField and
SetProcedureEntryPoint. This is problematic for removing unused
record-likes as the initialisation steps seem to "use" the value
produced by the record-like constructor even if the constructed cell
itself is never used.

Instead have a InitRecord step that takes the initial values for all of
the fields and and InitProcedure that additionally takes an entry point.
The entire step can be discarded if the record or procedure is not
used. This is especially useful for getting rid of stray closures that
are constructed for code that's later optimised out or inlined.