compiler: Use initialiser type for recursive vars
If a recursive value is immutable then we can look at the initialiser
expression's Scheme type when creating the recursive value. For lambdas
in particular this allows us to at least determine the arity of the
lambda without explicit type annotations. This is helpful for both error
reporting and optimising.

It's still preferred that recursive values use type annotations but this
gets us some of the advantages for free with untyped code.