stdlib: Add datum reader/writer fuzz driver
This reads a sequence of Scheme data from stdin and attempts to parse
it. If a read error or UTF-8 error is encountered then parsing will
resume at the next character. If another exception is raised, e.g. by
the C++ standard library, then execution will abort

For every successful parse the datum will be re-rendered to an internal
buffer and then reparsed. The reparsed datum is then tested for equality
to ensure that it was rendered correctly. If any exception is raised
while reparsing then execution will abort

This is intended for use by automated fuzzing tools such as afl-fuzz to
detect corner cases in datum parsing and rendering.