runtime: Explicitly render bytevectors in decimal
This fixes rendering a bytevector after rendering a character. The
character will leak the hex bit in to the the stream and then
renderBytevector will accidentally render its contents in hex

We might want to save and restore the ostream flags so we don't leak the
hex bit either but that's more work with no immediate benefit.