runtime: Fix reader EOF while consuming literal
If we encounter an EOF during consumeLiteral we should put back any
characters we've consumed so we can optionally re-parse the remaining
characters.

The test case added is the symbol "-i" which is attempted to be consumed
as "-inf.0". If we don't put the "-i" back then it cannot be properly
parsed.