compiler: Dispose unused vector allocations
The planner would previously initialise a vector by creating partly
initialised vector and then setting each of its elements. The setting
steps would "use" both the vector and its elements meaning neither could
be disposed by the planner even if the resulting vector was unused.

Instead, have InitVector take a vector of elements to create the vector
with and introduce a new InitFilledVector step that creates the vector
with a fill. As these take the new elements as input and produce the
initialised vector as output we can now dispose unused vectors properly.

At the same time we can also remove a trick in ForAll where we tried to
simplify generated IR for the types of loops that filled vector
initialisation was using.