compiler: Replace vector types with iv.KnownVector
Vector types were essentially unused except by the planner to determine
if a vector access is in-range and therefore capable of being inlined.
However, this was at a cost over approximately a thousand lines of code
and great complexity. Instead we can just have a KnownVector trait for
vectors that have a known length with much less code.

One case where this results is worse planning is indexing a constant
vector by a dynamic index. Previously the polymorphic declaration of
(vector-ref) would automatically determine the common type of the vector
and annotate the (vector-ref) result. This isn't enough to justify the
complexity of vector types.

Typed vectors are still a desired feature but they should have the
following properties the previous implementation didn't:

- Typed vectors should be created explicitly with its member type

- They should not be variant - e.g. (TypedVectorof <string>) should not
  be a subtype of (TypedVectorof <any>). This could be relaxed to allow
  read-only views in to vector "supertypes"

- They shouldn't allow dynamic type predicates. This would require
  encoding the member type information at runtime which would be
  exceedingly difficult