compiler: Basic pattern matching
This adds a new (match) primitive which takes a value to match and a
series of clauses. Each clause consists of a pattern datum followed by
zero or more body expressions.

The match clauses are evaluated in order until a pattern matches the
value. The clause's body is then evaluated with any new bindings
introduced by captures in the pattern. If no clause matches then a
match-error is signalled with the value as the evidence.

Currently the following patterns are supported:

- An underscore symbol will match any value and introduce no new bindings

- Other symbols will match any value and introduce that value as a new
  binding

- Self evaluating or quoted data will be tested for equality using
  (equals?)

- (cons) will match any pair with the given car and cdr patterns

- (list) will match any proper list with the given element patterns

- (vector) will match any vector with the given element patterns

- Record type constructors will pattern match the record type's fields
  in the order they appears as arguments for the constructor. Fields
  that do not appear in the constructor cannot be pattern matched.