compiler: Make special storage locations immutable
This makes storage locations defined using (define-report-procedure) and
record type constructors immutable. This is because the storage location
itself is annotated as special and mutating the binding could confuse
the compiler in to still using the special behaviour for the new value.
For example, this code would still work even though cons-rec-type has
been rebound:

(define-record-type <rec-type> (cons-rec-type) rec-type?)
(define rec-type-instance (cons-rec-type))
(set! cons-rec-type (lambda ()))

(match foo
  ((cons-rec-type rec-type-instance) #t))

This forcing record type constructors to be immutable is a violation of
R7RS but it's unlikely to be used in sane code.