More dense string and symbol representation
This changes string-like cells to contain an 8bit byte length and 8bit
character length. If the byte length isn't 255 then the rest of the
string-like cell is its inline data. Otherwise, its followed by a set of
32bit lengths and a pointer to its heap data.

This combines the large (20 byte) inline size of symbols with the higher
limits (32bit char/byte length) of strings.

One downside is that it becomes significantly more complicated to load
the byte length of a symbol. This functionality is used by
PlanSymbolEquality which uses it to optimise equality with symbol union
types. In the worst case we need to load the inline length and
conditionally load the heap length based on that.

The ability to plan (string-length) as a char length load has been
removed completely; the stdlib can be used instead.