compiler: Don't specify type for pattern clauses
When we bind a procedure we first plan it with an automatically
determined type and then cast it to the type of the storage location.
This type is typically <any> in which case no cast is required. However,
if the storage location has a type that isn't a strict subtype of the
automatically determined type when need to cast the procedure value.
This creates a thunk and interferes with inlining.

This should be fixed but for now we can avoid it in the pattern matching
clauses by not specifying a type for their storage location.