compiler: Properly calculate size of shadow stack
We were attempting to calculate the size of the shadow stack entry in
number of pointers based on the architecture. This was correct for
x86-32 where 64bit values are 32bit aligned. However, on 32bit ARM its
incorrect as it uses natural alignment for 64bit values.

Instead of tweaking the value further actually calculate the required
size using gelementptr.