compiler: Include CXXFLAGS and target information
This isn't very useful on x86-{32,64} where the compiler flags and
target mostly have an effect on CPU tuning for us. However, on ARM32
there are various incompatible ABI and FPU options that need to be
reconciled. This ensures these options are aligned between the compiler
and runtime using the standard CFLAGS/CXXFLAGS mechanism.

- Extract our target triple and data layout using an example clang++
  invocation when building the runtime. This ensures the runtime and
  compiler are targeting the same ABI and layout

- Include our CXXFLAGS in "required-clang-flags". This allows any
  platform specific flags to be exposed during final linking Llambda
  programs