runtime: Don't cache global constants when cloning
cachedClone() wasn't checking for global constants and instead was
allowing uncachedClone() to perform the check. This means when a
non-constant referenced a constant the constant (but not its children)
would be placed in the cloned cell cache. This is wasteful especially as
the cloned cell cache appears very high in actor profiles.

Instead hoist is the isGlobalConstant() check in to both uncachedClone()
and cloneCell(). This ensures global constants are never cached.`