Fix use of uninitialised memory
The cell allocator was using calloc() to allocate zero initialised
memory space. This is a legacy from when cell space was directly
allocated using mmap() which should always zero memory. Other parts of
the code were spending on zero filled space to avoid explicitly
initialising fields to zero, particularly the gcState field.

On most platforms we were finding the actual allocated size of the
calloc() to efficiently use any slack allocation space. However, on Mac
OS X at least calloc() will only zero the explicitly requested space.
This typically works until after the first GC cycle as new memory will
be allocated using mmap() or brk(). However, after the finalizer runs
it's likely non-zero memory will be released back to the malloc() pool.

This is particularly tricky as Valgrind indicates the usable size of an
allocation is the exact requested size. This makes the problem disappear
when run under Valgrind.

Fix by removing the assumptions in the code depending on zero filled
memory and using malloc(). This doesn't appear to have a large
performance impact and has the advantage of allowing Valgrind to check
for uninitialised access within the cell allocated area.