compiler: Special planning for (values) trampoline
(values) is used in idiomatic Scheme as an identity function. For
example, this can be used to generate the sequence of non-negative
integers:

(list-tabulate count values)

The native (values) implementation takes a rest argument list which is
wasteful when returning a single value as in above. A rest argument list
will be constructed and then the value will immediately be removed from
the list head.

Instead, add a specialised PlanIdentityProcedure which will plan a
procedure taking values and convert it to a return type. This can be
used to plan any identity procedure but is only reachable when planning
a trampoline for (values).

In the (list-tabulate) case this results in a nearly 2x speed up