compiler: Make InitPair discardable
InitPair was using the old pattern of having separate allocation
(InitPair) and set (SetPairC{a,d}r) steps. This makes the step
non-discardable as the set steps depend on the init step. The end result
of this is code like this can't be properly discarded:

(define x (computed-value)
(cons x x)

Instead have InitPair take its initial car and cdr values and mark the
step as discardable.