compiler: Make noreturn procedures return (U)
In C and C++ it's possible for functions to have non-vpid return types
but be marked as noreturn. This may have an affect on the calling
convention for the platform.

However, on the Scheme side its more accurate if we change the return
type to the empty Scheme type - (U). This especially important when
resolving polymorphic types as the <unit> return type from a procedure
that does not return should not contribute to the reconciled type
variable.