runtime: Merge ALWAYS_GC and NO_ADDR_REUSE
ALWAYS_GC and NO_ADDR_REUSE must be used together or else we terminate
segments incorrectly. Instead of fixing this just merge the two together
to reduce the size of the configuration space.

This means that 32bit builds will only check for leaks when GC debugging
is enabled. As it's very unlikely to have a 32bit only GC bug this isn't
a large loss. Also, ALWAYS_GC can still be manually even for 32bit
builds.