compiler: Fix SatisfiesSignature for void returns
SatisfiesSignature was not taking in to account that returns without a
representation type (i.e. <unit>) do not satisfy signatures that do
(i.e. <any>). This is because return types without a representation
return void so attempting to use their return value will read undefined
data off of the stack and possibly crash.