runtime: In-place updates of internal hash trees
When we're creating a new subtree or building from an assoc list we
don't need to preserve the original version of the tree we're working
on. When can use this to avoid copying array nodes by updating them
in-place instead. This approximately halves the number of allocations we
need to build a 800 entry hash tree.