compiler: Add optionalArgTypes to ProcedureType
This splits the current fixedArgTypes in to mandatoryArgTypes and
optionalArgTypes. Optional arguments have default values and appear
between the mandatory and rest arguments. This provides only initial
type system support.

This is inaccessible and currently always Nil. Where it was trivial to
handle the code has been adapted to consider types with optional
arguments but its mostly stubbed out.