compiler: Support parsing formals with optionals
This adds support for optional args to ParseFormals and adds a dedicated
ParseFormalsSuite test. The accepted forms are [arg default] and [arg :
<type> default]. This is based on the forms found in Typed Racket.

This also ports ExtractLambda to use the new optional arguments but it
has only been confirmed not to break the existing tests. The next
commits should test this explicitly and fix any breakage.