compiler: Remove polymorphic proc type shorthand
The shorthand was intended to be a copy of the Typed Racket shorthand.
However, we use:

  (All (A) A A R)

While Typed Racket uses an infix form:

  (All (A) A A -> R)

Our shorthand form is ambiguous - a longhand form of a procedure can
also be interpreted as a shorthand for a procedure accepting no
arguments and returning a procedure. This gets more hairy when we
support (->*) for optional types.