compiler: Optimisation for constant parameters
If (parameterize) is never used we can assume that all parameter
procedures without converters return their initial value. This means we
can either return the initial value directly (if its in scope or
constant) or annotate the type otherwise.

This is intended primarily to generate more efficient code for
input/output procedures when the input/output ports are not
parameterized.