compiler: Fix SatisfiesSignature for var args
Before the introduction of optional arguments we would correctly
consider a procedure without a rest arg being incompatible with one with
a rest arg. This is because the rest argument causes the native function
to have a different signature to accept the argument.

Switching to optional arguments changed the logic to be more abstract
and only consider the types of the var arg lists, not its presence. Fix
by explicitly testing for presence before types.

This fixes bad code generation in the remove-rxrs branch:wq