Remove (with-exception-handler)
In R7RS this seems to be mostly used as a primitive from which (guard)
can be built. To avoid the overhead of (call/cc) Llambda implemented the
core of (guard) directly in C++.

Remove as this is duplicate functionality and isn't usable sanely
without (call/cc) which is being removed.