compiler: Remove inline planning for (length)
Our inline (length) uses CalcProperListLength which doesn't use list
length hints. This means it will always scan the entire list.

The only other use of CalcProperListLength is for (case-lambda) dispatch
which should work with short lists. In that case the length hint is of
less value and the function call overhead likely makes it a net loss.