Remove (max)/(min) casting to flonum
R7RS defines that (max) and (min) return an inexact value if any of the
arguments are inexact. This makes sense in terms on the exact/inexact
distinction but makes less sense when those concepts are replaced with
an integer/flonum distinction.

This has a side benefit of avoiding a cell allocation in certain cases.