compiler: Add nocapture attribute for procedures
When specified this indicates that the procedure or native function does
not capture any of its pointer arguments. This corresponds and converts
to the LLVM "nocapture" parameter attribute for each parameter.

It is doubtful that LLVM can use this for any meaningful optimisations
as anything allocated from the garbage collector heap will already be
considered captured. However, this provides extremely valuable hints for
our upcoming escape analysis pass about which parameters do not escape.

Additionally, once our escape analysis can generate stack allocations
the LLVM optimiser will be able to perform optimisations on the basis
that those allocations are uncaptured.