compiler: Track the value of parameters in state
We previously had a hack for code which does not use `(parameterize)` at
all to track the value of parameters during execution. However, we can
also track the value of parameters within the lexical scope of a
`(make-parameter)` or `(parameterize)`. This works even in the face of
inlining which can greatly improve code generation for a lot of common
parameter use cases.