compiler: Move GC root determination to codegen
Previously the planner would indicate which TempValues needed to be GC
rooted using the `TempValue.isGcManaged` property. This had a number of
downsides:

- The planner is significantly larger than codegen so more code needed
  to be aware of GC values

- The rest of the GC logic lives in codegen; except for planning cell
  allocations the planner is nearly GC agnostic

- The planner was determining if a value should be GC rooted based on
  type instead of where the value came from. For example, casting a
  <boolean> to <any> would GC root the <any> even though it pointed to a
  constant.

- The conniver couldn't "un-root" values while rewriting steps without
  renaming to a new TempValue. This is a particular problem when stack
  allocating integers, chars, and flonums as they do not need to be GC
  rooted when they're allocated from the stack.

Fix by introducing a CollectableIrValue type in codegen that tracks an
IrValue and if it needs to be GC rooted. An implicit conversion makes
most of the existing code that expects a bare IrValue to continue
working.

Besides simpler code this reduces the number of excess GC roots required
in many common code patterns.